Crate vm_control

Handles IPC for controlling the main VM process.

The VM Control IPC protocol is synchronous, meaning that each VmRequest sent over a connection will receive a VmResponse for that request next time data is received over that connection.

The wire message format is a little-endian C-struct of fixed size, along with a file descriptor if the request type expects one.

Re-exports

pub use crate::gdb::*;

Modules

client

gdb

Structs

BalloonStats

BatControl
Used for VM to control battery properties.

GpuMemoryDesc
Struct that describes a GPU memory allocation that consists of up to 3 planes.

GpuMemoryPlaneDesc
Struct that describes the offset and stride of a plane located in GPU memory.

UsbControlAttachedDevice

Enums

BalloonControlCommand

BalloonControlResult

BalloonTubeCommand

BalloonTubeResult

BatControlCommand

BatControlResult

BatHealth

BatProperty

BatStatus

BatteryType

DiskControlCommand

DiskControlResult

FsMappingRequest

IrqSetup
Data to set up an IRQ event or IRQ route on the IRQ chip. VmIrqRequest::execute can't take an IrqChip argument, because of a dependency cycle between devices and vm_control, so it takes a Fn that processes an IrqSetup.

UsbControlCommand

UsbControlResult

VcpuControl
Control the state of a particular VM CPU.

VmIrqRequest

VmIrqResponse

VmMemoryRequest

VmMemoryResponse

VmMsyncRequest

VmMsyncResponse

VmRequest
A request to the main process to perform some operation on the VM.

VmResponse
Indication of success or failure of a VmRequest.

VmRunMode
Mode of execution for the VM.

Constants

USB_CONTROL_MAX_PORTS
The maximum number of devices that can be listed in one UsbControlCommand.

Type Definitions

MemSlot
An index in the list of guest-mapped memory regions.