Enum vm_control::VmMemoryRequest

pub enum VmMemoryRequest {
    RegisterMemory(SharedMemory),
    RegisterFdAtPciBarOffset(Alloc, SafeDescriptor, usize, u64),
    RegisterHostPointerAtPciBarOffset(Alloc, u64),
    RegisterVulkanMemoryAtPciBarOffset {
        alloc: Alloc,
        descriptor: SafeDescriptor,
        handle_type: u32,
        memory_idx: u32,
        physical_device_idx: u32,
        offset: u64,
        size: u64,
    },
    UnregisterMemory(MemSlot),
    AllocateAndRegisterGpuMemory {
        width: u32,
        height: u32,
        format: u32,
    },
    RegisterMmapMemory {
        descriptor: SafeDescriptor,
        size: usize,
        offset: u64,
        gpa: u64,
    },
}

Variants

RegisterMemory(SharedMemory)

Register shared memory represented by the given descriptor into guest address space. The response variant is VmResponse::RegisterMemory.

Tuple Fields of RegisterMemory

0: SharedMemory

RegisterFdAtPciBarOffset(Alloc, SafeDescriptor, usize, u64)

Similiar to VmMemoryRequest::RegisterMemory, but doesn’t allocate new address space. Useful for cases where the address space is already allocated (PCI regions).

Tuple Fields of RegisterFdAtPciBarOffset

0: Alloc1: SafeDescriptor2: usize3: u64

RegisterHostPointerAtPciBarOffset(Alloc, u64)

Similar to RegisterFdAtPciBarOffset, but is for buffers in the current address space.

Tuple Fields of RegisterHostPointerAtPciBarOffset

0: Alloc1: u64

RegisterVulkanMemoryAtPciBarOffset

Similiar to RegisterFdAtPciBarOffset, but uses Vulkano to map the resource instead of the mmap system call.

Fields of RegisterVulkanMemoryAtPciBarOffset

alloc: Allocdescriptor: SafeDescriptorhandle_type: u32memory_idx: u32physical_device_idx: u32offset: u64size: u64

UnregisterMemory(MemSlot)

Unregister the given memory slot that was previously registered with RegisterMemory*.

Tuple Fields of UnregisterMemory

0: MemSlot

AllocateAndRegisterGpuMemory

Allocate GPU buffer of a given size/format and register the memory into guest address space. The response variant is VmResponse::AllocateAndRegisterGpuMemory

Fields of AllocateAndRegisterGpuMemory

width: u32height: u32format: u32

RegisterMmapMemory

Register mmaped memory into the hypervisor’s EPT.

Fields of RegisterMmapMemory

descriptor: SafeDescriptorsize: usizeoffset: u64gpa: u64

Implementations

impl VmMemoryRequest

pub fn execute(
self,
vm: &mut impl Vm,
sys_allocator: &mut SystemAllocator,
map_request: Arc<Mutex<Option<ExternalMapping>>>,
gralloc: &mut RutabagaGralloc
) -> VmMemoryResponse

Executes this request on the given Vm.

Arguments

vm - The Vm to perform the request on.

allocator - Used to allocate addresses.

This does not return a result, instead encapsulating the success or failure in a VmMemoryResponse with the intended purpose of sending the response back over the socket that received this VmMemoryResponse.